Enable releases from CI to the test repo (#15)
* Removed unnecessary task

One can just inspect the "dist" dir. Previously it was useful because we had the "dist" filtering logic implemented in Gradle build script. Now it has been pushed down to SBT.

* Enable release from CI

We're ready. If this PR is merged we will release to the 'test-repo'. This way we can test if everything works. If it does, we should clear the last TODO and change the repo to 'maven'